Add coverage.py.
This will help find the parts of the app that need tests (which is
most of it).

TO RUN COVERAGE:

`coverage run ./manage.py test app`,

where `app` is probably `taxi` or `mongologin`. This will create a
.coverage file in the current directory. Running `coverage report` now
will give you a nice report in the terminal indicating how many Python
statements were actually run during the test and how many are missing.

"But wait", you say, "all those percentages are a bit meaningless. I
want to know which lines to test!"

Well, you're in luck! Running `coverage html` will generate an
`htmlcov` directory. Open up `htmlcov/index.html` and you'll get a
pretty swanky report of which lines are covered (green) and which ones
are missing (red). Make all the green turn to red and your code will
work.